Fix slog.With() and slog.Group() not working with the local handler
I had to rewrite the local handler using a custom writer because there
wasn't an easy way to override the output of either JSONHandler or
TextHandler in slog.

The bug that presented it self was that the attrs from the parent logger
was missing when creating child loggers with slog.With() and
slog.Group().